Use Werkzeug's context locals to store request metadata
This fixes an issue (and an ugly architecture debt) that caused the APM agent to lose track of the endpoint name when multiple requests were processed at the same time. The endpoint name would "leak" on the previous request, and the next request would not have the endpoint name saved.